multiwoz_env/requirements.txt
absl-py==2.3.1
accelerate==1.6.0
ace_tools==0.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250224
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
attrs==25.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.48.2
blake3==1.0.4
bleach==6.2.0
Brotli==1.1.0
cachetools==5.5.2
cbor2==5.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.1
compressed-tensors==0.11.0
contourpy==1.3.2
cryptography==39.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.11
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
execnet==2.1.1
executing==2.2.0
faiss-cpu==1.11.0
Faker==24.14.1
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fastapi-login==1.9.3
fastjsonschema==2.20.0
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
freezegun==1.2.2
frozendict==2.4.7
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
gguf==0.16.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==3.2.1
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.34.1
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
interegular==0.3.3
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.0
json5==0.9.25
jsonnet==0.20.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.2.2
libclang==18.1.1
libcst==1.7.0
litellm==1.37.19
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.11.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mistune==2.0.4
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
munch==4.0.0
namex==0.1.0
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.8.0
openai-harmony==0.0.8
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.7
opt_einsum==3.4.0
optree==0.17.0
orjson==3.10.18
outcome==1.3.0.post0
outlines_core==0.2.11
overrides==7.4.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pendulum==3.1.0
pexpect==4.9.0
pillow==10.4.0
pip==25.1
platformdirs==4.3.7
pluggy==1.5.0
polyfactory==2.21.0
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.7
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.51.1
recoma==0.0.5
referencing==0.36.2
regex==2024.11.6
registrable==0.0.4
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.24.0
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.5
soxr==1.0.0
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.10
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.17.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.22.1
tomlkit==0.13.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
treelib==1.7.1
trio==0.21.0
triton==3.4.0
trl==0.17.0
typer==0.20.0
types-python-dateutil==2.9.0.20250708
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.27.1
uvloop==0.19.0
vllm==0.11.0
wandb==0.19.11
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
